Added holding the main window handle into the mantis client for aligning rendering if the window is moved.
Added JS C++ interop. The first prototypes that are available is setElementContent and setElementDisplay for setting the innerHTML in javascript and the style.display.

Also has the ability for any element in HTML to call app.onMethodCall(args); and have them parsed on the C++ side to call various in-engine functions (think connect, boot, set armor etc)